console: move "reopen_copyedit" function to its own URL.
When a project has been submitted, accepted, and copyedited, it is in
the NEEDS_APPROVAL stage.  During this stage, the "awaiting_authors"
console page is used to view and manage the project.  In particular,
the editor may click the 'reopen_copyedit' button to reopen the
project.

We can simplify the logic by moving the form submission handling into
another view, changing the form submission URL.

(This page is only accessible to editors, is rarely used, and the form
is trivial, so there's no need for backward compatibility here.)