ActiveProject: allow reopening copyedit after approval.
If a project has already been approved by all authors (so is in the
NEEDS_PUBLICATION state) and last-minute changes are needed, we still
want to be able to reopen copyediting.

If a project is in NEEDS_PUBLICATION state, then
self.author_approval_datetime would have been set (by approve_author)
at the time the final author approved.  When reverting to
NEEDS_COPYEDIT state, this field should be reverted accordingly.