console: allow reopening copyedit from the Publish Submission page.
If a project has already been approved by all authors (so is in the
NEEDS_PUBLICATION state) and last-minute changes are needed, we still
want to be able to reopen copyediting.

Move the reopen_copyedit form ("card" and "modal") into a separate
template which is included in both pages; and allow this form to be
submitted if the project is in NEEDS_PUBLICATION state.