CoreProject, UnpublishedProject: new method future_doi_prefix.
This method returns the prefix of the DOI that will be assigned for
the project, if and when the project is published.  If a DOI has
already been registered for this project, use that prefix; otherwise
use the configured prefix that will be used at registration time.

This is meant to be used in project previews, to show the format of
recommended citations without showing the actual DOI in a way that
might suggest the content is final and citable.