project_preview: use the correct future DOI prefix.
In previews, display the appropriate prefix for placeholder DOIs.  If
the site is not a DOI issuer and the project doesn't have a
manually-assigned DOI, then don't display DOIs at all.

(Continue to display a placeholder even if the project actually has an
assigned DOI; we don't want to give the impression that the content is
final and citable.)